backend/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flask==3.0.1
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
werkzeug==3.0.1
zipp==3.17.0
requests==2.31.0
python-dotenv
flask_cors
pytest
gunicorn
pymongo
openai
langchain
bs4
